==> 02_Second_Scraping_Notebook.ipynb <==
Imports

import praw
import json
import time
import pandas as pd

Load in credentials / initiate PRAW

json_file = open('Reddit_creds.json')
json_str = json_file.read()
json_data = json.loads(json_str)

reddit = praw.Reddit(user_agent=json_data['user_agent'],
                     client_id=json_data['client_id'], client_secret=json_data['client_secret'])

Define subreddits to scrape

yang_sub = reddit.subreddit('YangForPresidentHQ')

pete_sub = reddit.subreddit('Pete_Buttigieg')

sanders_sub = reddit.subreddit('SandersForPresident')

harris_sub = reddit.subreddit('Kamala')

biden_sub = reddit.subreddit('JoeBiden')

#Adding democratic subs so I can include users who are democrats but aren't posting on one of the 5 candiate subreddits
dem_sub_1 = reddit.subreddit('centerleftpolitics')

dem_sub_2 = reddit.subreddit('VoteBlue')

dem_sub_3 = reddit.subreddit('BlueMidterm2018')

dem_sub_4 = reddit.subreddit('liberal')

dem_sub_5 = reddit.subreddit('democrats')

subs = [yang_sub,pete_sub,sanders_sub,harris_sub,biden_sub,dem_sub_1,dem_sub_2,dem_sub_3,dem_sub_4,dem_sub_5]

Create a list of users from recent comments on these subs

users = []

#For each sub above, find the 1,000 most recent comments and append the comment author onto the users list
#convert the list to a set to remove duplicates and then back to a list
start = time.time()
for sub in subs:
    for comment in sub.comments(limit=1000):
        users.append(comment.author)
        
print(len(users))
users = list(set(users))
print(len(users))
end = time.time()
print(end-start)

Actual Scraping Code

# For each user in users, scrape their most recent 1,000 comments
# Scrape run on 8/17/19

start = time.time()
text = []
author = []
subreddit = []
postid = []
counter = 0
for user in users:
    if counter % 10 == 0:
        prog = time.time()
        print(counter)
        print((prog-start)/60)
    try:  #handle cases where redditor is deleted or something
        for comment in user.comments.new(limit=1000):
            text.append(comment.body)
            author.append(comment.author)
            postid.append(comment.id)
            subreddit.append(comment.subreddit.display_name)
    except:
        pass
    counter += 1
end = time.time()
print(end-start)

Exporting results

#save scrape data as a dataframe
df = pd.DataFrame([text,author,subreddit,postid],\
                        index=['Comment Text','Author','Subreddit','Post ID']).T

df.shape

#export scrape info to a .csv
df.to_csv('full_scrape_3_add.csv',index=False)
