==> 03_Combine_Scrapes.ipynb <==
#Import Pandas
import pandas as pd

#Load in 1st scrape
df_comb = pd.read_csv('combined_scrape.csv')

df_comb.head()

#Load in 2nd scrape
df_full_3 = pd.read_csv('full_scrape_3_add.csv')

df_full_3.head()

df_comb.shape

df_full_3.shape

#Combine Dataframes
df = pd.concat([df_comb,df_full_3],axis=0)

df.shape

#Drop duplicate comments
df.drop_duplicates(subset='Post ID',inplace=True)

df.shape

#Save to CSV
df.to_csv('combined_scrape_2.csv',index=False)
