==> 04_Create_Target_Column.ipynb <==
Imports

import pandas as pd
from sklearn.model_selection import train_test_split

Load in Dataframe

df = pd.read_csv('combined_scrape_2.csv')

df.shape

df.head()

df.dtypes

Create Target Column

#Group the data by author and subreddit
sum_series = df.groupby(['Author','Subreddit']).size()

#Create a list of users in my initial dataframe
users = list(set([x for x in sum_series.index.get_level_values(0)]))

len(users)

# For the political subreddits that I want to define my target, identify the number of posts a user made in each
# subreddit and save to a dictionary
political_dict={}
for user in users:
    inner_dict = {}
    try:
        inner_dict['yang'] = sum_series[user]['YangForPresidentHQ']
    except:
        inner_dict['yang'] = 0
    try:
        inner_dict['bernie'] = sum_series[user]['SandersForPresident']
    except:
        inner_dict['bernie'] = 0
    try:
        inner_dict['bernie'] += sum_series[user]['wayofthebern']
    except:
        inner_dict['bernie'] += 0
    try:
        inner_dict['bernie'] += sum_series[user]['Political_Revolution']
    except:
        inner_dict['bernie'] += 0
    try:
        inner_dict['pete'] = sum_series[user]['Pete_Buttigieg']
    except:
        inner_dict['pete'] = 0
    try:
        inner_dict['kamala'] = sum_series[user]['Kamala']
    except:
        inner_dict['kamala'] = 0
    try:
        inner_dict['biden'] = sum_series[user]['JoeBiden']
    except:
        inner_dict['biden'] = 0
    try:
        inner_dict['warren'] = sum_series[user]['ElizabethWarren']
    except:
        inner_dict['warren'] = 0
    try:
        inner_dict['tulsi'] = sum_series[user]['tulsi']
    except:
        inner_dict['tulsi'] = 0
    try:
        inner_dict['dem'] += sum_series[user]['centerleftpolitics']
    except:
        inner_dict['dem'] = 0
    try:
        inner_dict['dem'] += sum_series[user]['VoteBlue']
    except:
        inner_dict['dem'] += 0
    try:
        inner_dict['dem'] = sum_series[user]['BlueMidterm2018']
    except:
        inner_dict['dem'] = 0
    try:
        inner_dict['dem'] = sum_series[user]['liberal']
    except:
        inner_dict['dem'] = 0
    try:
        inner_dict['dem'] = sum_series[user]['democrats']
    except:
        inner_dict['dem'] = 0
    political_dict[user]=inner_dict


# Create a dataframe from my dictionary
df_target = pd.DataFrame(political_dict).T

df_target.head(5)

# Define a user as a Yang user
# if they have 5 or more posts on the subreddits in question, this is used to exclude 
## people who might not be democrats or a supporter of a 2020 democratic candiate
# if they've posted more in the yang subreddit than any of the other candidate subreddits

def func(row):
    total = sum(row)
    if total < 5:
        return 'drop'
    if (row['yang'] > row['bernie']) & (row['yang'] > row['tulsi']) & (row['yang'] > row['warren']) & (row['yang'] > row['biden']) & (row['yang'] > row['kamala']) & (row['yang'] > row['pete']):
        return '1' #1 = yang supporter
    else:
        return '0' #0 = not yang supporter

# Apply function to create target column
df_target['target'] = df_target.apply(func,axis=1)

df_target['target'].value_counts()

df_target.drop(['bernie','biden','kamala','pete','yang','tulsi','warren','dem'],axis=1,inplace=True)

df_target = df_target[df_target['target']!='drop']

df_target.head()

# Train test split my users based on the target column
y_train,y_test = train_test_split(df_target['target'],random_state=42,stratify = df_target['target'])

y_train.size

y_test.size

Export dataframes of train users and test users

pd.DataFrame(y_train).to_csv('fin_users_train.csv')

pd.DataFrame(y_test).to_csv('fin_users_test.csv')
