==> 05_Subreddit_Group.ipynb <==
Imports

import pandas as pd

Load in Dataframes

df = pd.read_csv('combined_scrape_2.csv')

df.head()

df_target_train = pd.read_csv('fin_users_train.csv')

df_target_train.head()

Remove non-train data

users = list(df_target_train['Unnamed: 0'])

#Simple function to check if comment author is in my list of training users
def clean_df(row):
    if row['Author'] in users:
        return '1'
    else:
        return '0'


df['keep'] = df.apply(clean_df,axis=1)

df.head()

df = df[df['keep']=='1']

df.shape

df.head()

Identify most common subreddits and export results

subs_df = pd.DataFrame(df.Subreddit.value_counts())

subs_df.head()

subs_df.shape

#Only keep most recent 1000 subreddits
subs_df_1000 = subs_df[subs_df['Subreddit']>=1000].copy()

subs_df_1000.shape

#Add number of unique users posting in each subreddit
user_count = []
for i in range(subs_df_1000.shape[0]):
    sub = subs_df_1000.iloc[i,].name
    user_count.append(len(df[df['Subreddit']==sub]['Author'].value_counts()))

subs_df_1000['users']=user_count

subs_df_1000.head()

#Export to csv
subs_df_1000.to_csv('fin_subreddit_1000plus.csv')
