==> 07_Remove_features_&_2nd_model.ipynb <==
Imports

import pandas as pd
import matplotlib.pyplot as plt
from imblearn.over_sampling import SMOTE
from sklearn.metrics import confusion_matrix, recall_score

Load in dataframes

df = pd.read_csv('combined_scrape_2.csv')

df.shape

df.head()

df_target_train = pd.read_csv('fin_users_train.csv')

df_target_train.shape

df_target_test = pd.read_csv('fin_users_test.csv')

df_target_test.shape

df_target = pd.concat([df_target_test,df_target_train])

df_target.shape

Create subreddit features

users = list(df_target['Unnamed: 0'])

def clean_df(row):
    if row['Author'] in users:
        return '1'
    else:
        return '0'


df['keep'] = df.apply(clean_df,axis=1)

df = df[df['keep']=='1']

#Group relevant posts by subreddit and author
sum_series = df.groupby(['Author','Subreddit']).size()

#Add in the subreddits I want to create features on based on my spreadsheet analysis
#Remove subreddits that I identified from my first model run as not predictive
list_of_subs = ['ENLIGHTENEDCENTRISM','Fuckthealtright','LateStageCapitalism','ABoringDystopia','nba',
                'CollegeBasketball','lakers','bostonceltics','sixers','Conservative','tuesday',
                'CryptoCurrency','Bitcoin','nfl','CFB','CHIBears','buildapcsales','Amd','Games',
                'gaming','conspiracy','neoliberal','stupidpol','Android','anime','askgaybros','atheism',
                'aznidentity','BasicIncome','BigBrother','books','cars','Christianity','Documentaries',
                'Economics','environment','Futurology','hiphopheads','hockey','investing','JoeRogan',
                'JordanPeterson','keto','liberalgunowners','Libertarian','marvelstudios','MMA','movies',
                'Music','rupaulsdragrace','samharris','StarWars','technology','teenagers','television',
                'teslamotors','thebachelor','TheLastAirbender','Tinder','trees','TwoXChromosomes',
                'cats','Cricket','Fitness','gratefuldead','HomeImprovement']

#Create a dictionary of user / subreddits
features = {}
for user in users:
    feature_dict = {}
    for sub in sum_series[user].index:
        if sub in list_of_subs:
            feature_dict[sub]=sum_series[user][sub]
    if len(feature_dict)!=0:
        features[user]=feature_dict


df_feature = pd.DataFrame(features).T

df_feature.head()

df_feature.shape

#NaNs at this stage are subs with no posts, replace NaN with "0"
df_feature.fillna(0,inplace=True)

#Combine related subs into a single feature
df_feature['anti_alt_right_comb']=df_feature['ENLIGHTENEDCENTRISM']+df_feature['Fuckthealtright']
df_feature['bad_capitalism_comb']=df_feature['LateStageCapitalism']+df_feature['ABoringDystopia']
df_feature['basketball_comb']=df_feature['nba']+df_feature['CollegeBasketball']+df_feature['lakers']+df_feature['bostonceltics']+df_feature['sixers']
df_feature['conservative_comb']=df_feature['Conservative']+df_feature['tuesday']
df_feature['crypto_comb']=df_feature['CryptoCurrency']+df_feature['Bitcoin']
df_feature['football_comb']=df_feature['nfl']+df_feature['CFB']+df_feature['CHIBears']
df_feature['pc_comb']=df_feature['buildapcsales']+df_feature['Amd']
df_feature['gaming_comb']=df_feature['gaming']+df_feature['Games']


#remove subreddits that were used in a combination feature
df_feature.drop(['ENLIGHTENEDCENTRISM','Fuckthealtright','LateStageCapitalism','ABoringDystopia',
                 'nba','CollegeBasketball','sixers','lakers','bostonceltics','Conservative','tuesday',
                 'CryptoCurrency','Bitcoin','nfl','CFB','CHIBears',
                 'buildapcsales','Amd','Games','gaming'],axis=1,inplace=True)

df_feature = df_feature.div(df_feature.sum(axis=1), axis=0)

df_feature.head(20)

df_feature.isnull().sum().head()

Split features dataframe into test and training ones

df_target_train = df_target_train.set_index('Unnamed: 0')

df_target_test = df_target_test.set_index('Unnamed: 0')

df_feature.shape

df_train = pd.concat([df_feature,df_target_train],axis=1,join='inner')

df_test = pd.concat([df_feature,df_target_test],axis=1,join='inner')

df_train.shape

df_test.shape

#Export out final train dataframe for additional models
df_train.to_csv('fin_train_df.csv')

#Export out final test dataframe for additional models
df_test.to_csv('fin_test_df.csv')

Second Logisitic Regression Classifier

from sklearn.linear_model import LogisticRegression

lr = LogisticRegression(solver='lbfgs')

X_train = df_train.drop('target',axis=1)

y_train = df_train['target']

X_test = df_test.drop('target',axis=1)

y_test = df_test['target']

lr.fit(X_train,y_train)

lr.score(X_train,y_train)

y_train.value_counts(normalize=True)

lr.score(X_test,y_test)

y_test.value_counts(normalize=True)

pd.Series(lr.predict(X_test)).value_counts(normalize=True)

confusion_matrix(y_test,lr.predict(X_test))

y_test.value_counts()

#Calculate sensitivity
recall_score(y_test,lr.predict(X_test))

LR model using SMOTE

smt = SMOTE(random_state=42)

X_train_2,y_train_2 = smt.fit_sample(X_train,y_train)

lrsmt = LogisticRegression(solver='lbfgs')

lrsmt.fit(X_train_2,y_train_2)

lrsmt.score(X_train_2,y_train_2)

lrsmt.score(X_test,y_test)

pd.Series(lrsmt.predict(X_test)).value_counts(normalize=True)

confusion_matrix(y_test,lrsmt.predict(X_test))

y_test.value_counts()

#Calculate sensitivity
recall_score(y_test,lrsmt.predict(X_test))

results = pd.DataFrame(lrsmt.coef_.T,X_test.columns)

#Identify features with the greatest predictive power
results[abs(results[0])>1].sort_values([0],ascending=False)

These results are both informative and interesting. The strongest subreddits or topics of interests associated with Yang supporters are "BasicIncome" (Yang's central policy proposal is universal basic income), "aznidentity" (Yang is Asian-American so it makes sense that other Asian Americans would identify with him) and "samharris" (One of Yang's first widespread exposure moments was on the Sam Harris podcast). The less expected and thus more informative results / areas are where current Yang supporters, especially ones who are already interested in the topic can find other potential Yang supporters. Some of these topics/subreddits are crypto currency, hip hop fans, and marijuana users (r/trees is a subreddit for marijuana enthusiasts)
