==> 08_Create_NLP_features.ipynb <==
Imports

import pandas as pd

Load in dataframes

df = pd.read_csv('combined_scrape_2.csv')

df.shape

df.head()

df.dtypes

df_yang_train = pd.read_csv('fin_users_train.csv')

df_yang_test = pd.read_csv('fin_users_test.csv')

df_yang_train.shape

df_yang_test.shape

2487+830

#Create a list of users from both train and test dataframe, goal here is to only use the comments for these users
## for my NLP features
users = list(df_yang_train['Unnamed: 0'])+list(df_yang_test['Unnamed: 0'])

len(users)

Extract out relevant comments

# Code to scrape just comments from the YangForPresidentHQ subreddit
# user_comment_dict = {}
# len = df.shape[0]
# per = len // 100
# for i in range(df.shape[0]):
#     if i % per ==0:
#         print(i//per)
#     if df.iloc[i,:]['Subreddit'] == 'YangForPresidentHQ':
#         author = df.iloc[i,:]['Author']
#         try:
#             user_comment_dict[author] = str(user_comment_dict.get(author,"")) + " " + df.iloc[i,:]['Comment Text']
#         except:
#             pass

# Code to scrape comments from just the politics subreddit
# Appends all of a users comments into one string
# Returns a number for every percent of the scrape that is done

user_comment_dict = {}
len = df.shape[0]
per = len // 100
for i in range(df.shape[0]):
    if i % per ==0:
        print(i//per)
    if df.iloc[i,:]['Subreddit'] == 'politics':
        author = df.iloc[i,:]['Author']
        try:
            user_comment_dict[author] = str(user_comment_dict.get(author,"")) + " " + df.iloc[i,:]['Comment Text']
        except:
            pass

pd.Series(user_comment_dict).to_csv('fin_comments_politics.csv',header=False)

# pd.Series(user_comment_dict).to_csv('yangorno_comments_yangsub.csv',header=False)
