==> 09_Process_&_Clean_NLP.ipynb <==
Imports

import pandas as pd
import nltk
from nltk.tokenize import RegexpTokenizer
import regex as re

Load in dataframe

df = pd.read_csv('fin_comments_politics.csv',names=['User','Text'])

df.head()

Parse/clean comment text

tokenizer = RegexpTokenizer('\s+', gaps=True)

#This function converts a string of raw text into a list of individual words (tokens)

def processText(raw_text):
    
    tokens = tokenizer.tokenize(raw_text) #Split text into word tokens on spaces
    
    token_new=[]
    
    for token in tokens:
        if 'http' in token: #remove web addresses
            pass
        else:
            result = re.sub("[^a-zA-Z]", " ",token.lower()) #replace all non-letters with a space
            tokens2 = tokenizer.tokenize(result) #if a token now has space, split into multiple tokens
            for token2 in tokens2:
                if (len(token2) > 1) & (token2 != 'xb'): #remove single letter tokens and the xb token
                    token_new.append(token2)
    
    return(" ".join(token_new))

df.iloc[6,:]['Text']

processText(df.iloc[6,:]['Text'])

df['Clean'] = df['Text'].apply(processText)

df.head()

# df.to_csv('fin_cleaned_nlp.csv')
