==> 10_NLP_model.ipynb <==
Imports

import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
from nltk.stem import WordNetLemmatizer
from nltk.corpus import stopwords 
from nltk.tokenize import RegexpTokenizer
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import confusion_matrix,recall_score
from imblearn.over_sampling import SMOTE

Load in dataframes

df_train = pd.read_csv('fin_users_train.csv')

df_train.head()

df_train = df_train.set_index('Unnamed: 0')

df_test = pd.read_csv('fin_users_test.csv')

df_test.head()

df_test = df_test.set_index('Unnamed: 0')

df_train.head()

df_nlp = pd.read_csv('fin_cleaned_nlp.csv')

df_nlp.head()

df_nlp.set_index('User',inplace=True)

df_nlp.isnull().sum()

df_nlp.dropna(axis=0,inplace=True)

Remove stopwords and lemmatize the NLP dataframe

tokenizer = RegexpTokenizer('\s+', gaps=True)
lemmatizer = WordNetLemmatizer()
stops = set(stopwords.words('english'))

#Add candidates to the stop words list since these will greatly skew results and not help us identify the topics they talk about
stops.update(['yang','andrew','pete','buttigieg','mayor','sanders','bernie','bern','biden','kamala','harris','warren','elizabeth','tulsi','gabbard'])

#don can be a nickname for donald and has relevance for this project
stops.remove('don')

#Actual function that tokenizes text, removes stop words, lemmatizes the tokens, and returns the tokens joined with a space
def stop_and_lemma(text):
    processed_tokens=[]
    tokens = tokenizer.tokenize(text)
    for token in tokens:
        if token not in stops:
            processed_tokens.append(lemmatizer.lemmatize(token))
    return(" ".join(processed_tokens))

df_nlp['Use'] = df_nlp['Clean'].apply(stop_and_lemma)

df_nlp.drop(labels=['Unnamed: 0','Text','Clean'],axis=1,inplace=True)

Create train and test dataframes

#Using inner joins here to remove out users who did not have a post in r/politics

df_train = pd.concat([df_train,df_nlp],join='inner',axis=1)
df_train['predicted_target']=df_train['target']
df_train.drop(labels=['target'],axis=1,inplace=True)

df_test = pd.concat([df_test,df_nlp],join='inner',axis=1)
df_test['predicted_target']=df_test['target']
df_test.drop(labels=['target'],axis=1,inplace=True)

Applying TF-IDF vectorizer, Logisitic Regression Classifier and hypertuning

#Due to the multiple operations that needed to be run on these dataframes between the vecotorizing and
## the logisitic regression, I opted to use a for loop to tune my hyperparameters as opposed to a grid search
for n_grams in [(1,1),(1,2),(1,3)]:
    for features in [100,500,1000]:
        #initialize TF-IDF vectorizer
        tf = TfidfVectorizer(ngram_range = n_grams,max_features=features)
        #Generate sparse matrixes and combine with previous features
        sparseXtrain = tf.fit_transform(df_train['Use'])
        sparseXtrain_df= pd.DataFrame(sparseXtrain.toarray(),
               columns=tf.get_feature_names(),index=df_train.index)
        df_train2 = pd.concat([df_train.drop('Use',axis=1),sparseXtrain_df],axis=1,join='inner')

        sparseXtest = tf.transform(df_test['Use'])
        sparseXtest_df= pd.DataFrame(sparseXtest.toarray(),
               columns=tf.get_feature_names(),index=df_test.index)
        df_test2 = pd.concat([df_test.drop('Use',axis=1),sparseXtest_df],axis=1,join='inner')

        #Create X_Train/y_train
        X_train = df_train2.drop('predicted_target',axis=1)
        X_test = df_test2.drop('predicted_target',axis=1)
        y_train = df_train2['predicted_target']
        y_test = df_test2['predicted_target']

        #Run logisitic Regression
        lr = LogisticRegression(solver='lbfgs')
        lr.fit(X_train,y_train)
        print('Scores For: ',n_grams,features)
        print(lr.score(X_train,y_train))
        print(lr.score(X_test,y_test))

#count vectorizer model generally doesn't converge, scores look worse than TF-IDF so not using

# for n_grams in [(1,1),(1,2)]:
#     for features in [100,500]:
#         #initialize Count Vectorizer
#         cv = CountVectorizer(ngram_range = n_grams,max_features=features)
#         #Generate sparse matrixes and combine with previous features
#         sparseXtrain = cv.fit_transform(df_train['Use'])
#         sparseXtrain_df= pd.DataFrame(sparseXtrain.toarray(),
#                columns=cv.get_feature_names(),index=df_train.index)
#         df_train2 = pd.concat([df_train.drop('Use',axis=1),sparseXtrain_df],axis=1,join='inner')

#         sparseXtest = cv.transform(df_test['Use'])
#         sparseXtest_df= pd.DataFrame(sparseXtest.toarray(),
#                columns=cv.get_feature_names(),index=df_test.index)
#         df_test2 = pd.concat([df_test.drop('Use',axis=1),sparseXtest_df],axis=1,join='inner')

#         #Create X_Train/y_train
#         X_train = df_train2.drop('predicted_target',axis=1)
#         X_test = df_test2.drop('predicted_target',axis=1)
#         y_train = df_train2['predicted_target']
#         y_test = df_test2['predicted_target']

#         #Run logisitic Regression
#         lr = LogisticRegression(solver='lbfgs')
#         lr.fit(X_train,y_train)
#         print(lr.score(X_test,y_test),"variables: ",n_grams,features)

#Using the TF-IDF vecotrizer that had the best score and ran in the shortest time
tf = TfidfVectorizer(ngram_range = (1,2),max_features=1000)
        
#Generate sparse matrixes and combine with previous features
sparseXtrain = tf.fit_transform(df_train['Use'])
sparseXtrain_df= pd.DataFrame(sparseXtrain.toarray(),
       columns=tf.get_feature_names(),index=df_train.index)
df_train2 = pd.concat([df_train.drop('Use',axis=1),sparseXtrain_df],axis=1,join='inner')

sparseXtest = tf.transform(df_test['Use'])
sparseXtest_df= pd.DataFrame(sparseXtest.toarray(),
       columns=tf.get_feature_names(),index=df_test.index)
df_test2 = pd.concat([df_test.drop('Use',axis=1),sparseXtest_df],axis=1,join='inner')

#Create X_Train/y_train
X_train = df_train2.drop('predicted_target',axis=1)
X_test = df_test2.drop('predicted_target',axis=1)
y_train = df_train2['predicted_target']
y_test = df_test2['predicted_target']

#Run logisitic Regression
lr = LogisticRegression(solver='lbfgs')
lr.fit(X_train,y_train)
print('Scores For: (1,2) & 1000 features')
print(lr.score(X_train,y_train))
print(lr.score(X_test,y_test))

confusion_matrix(y_test,lr.predict(X_test))

y_test.value_counts()

y_test.value_counts(normalize=True)

recall_score(y_test,lr.predict(X_test))

#Unbalanced classes are causing us to underpreict on the positive class, let's use an oversampling technique

Apply Oversampling of Minority Class

smt = SMOTE(random_state=42)

X_train_bal,y_train_bal = smt.fit_sample(X_train,y_train)

lrsmt = LogisticRegression(solver='lbfgs')

lrsmt.fit(X_train_bal,y_train_bal)

lrsmt.score(X_train_bal,y_train_bal)

lrsmt.score(X_test,y_test)

confusion_matrix(y_test,lrsmt.predict(X_test))

recall_score(y_test,lrsmt.predict(X_test))

#Resulting dataframe showing words with the most predictive power
pd.DataFrame(lrsmt.coef_.T,X_train.columns).sort_values(by=0,ascending=False)

These results aren't that surprising. The strongest word associated with Yang supporters is "ubi" which stands for universal basic income and is the center topic in Andrew Yang's platform. Other highly associated words are "welfare" (perhaps indicating Yang supporters discussing how ubi would interact with current welfare platforms or how ubi isn't the same as welfare), "debate" (a big topic of Yang supporters the last couple months are how he did in the July debates and if he'll qualify for the Sept/Oct debates), and "freedom" (Andrew's ubi proposal is known as the "freedom dividend"). Words that are not associated with Yang supporters are a bit more interesting and harder to explain. The word that's least associated with Yang supporters is "f&ast;&ast;&ast;". One guess for this is "f&ast;&ast;&ast;" is often used in conjunction with Trump. Many supporters of Yang are known to be more moderate or even previous Trump supporters so one would expect Yang supporters to probably be critical of the president and his politics but less outright hostile and willing to drop blanket anti-Trump phrases such as "f&ast;&ast;&ast; Trump" or "f&ast;&ast;&ast; republicans". Another interesting word here is "corporate", perhaps indicating that Yang supporters are less likely to complain about corporations and corporate influence in politics than non-Yang supporters.
